+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/export/_base.py
@@ -43,6 +43,7 @@ from azure.monitor.opentelemetry.exporter._constants import (
     _RETRYABLE_STATUS_CODES,
     _THROTTLE_STATUS_CODES,
     DropCode,
+    _CLIENT_EXCEPTION,
 )
 # from azure.monitor.opentelemetry.exporter._configuration import _ConfigurationManager
 from azure.monitor.opentelemetry.exporter._connection_string_parser import ConnectionStringParser
@@ -341,7 +342,7 @@ class BaseExporter:
                         else:
                             if not self._is_stats_exporter():
                                 if self._customer_sdkstats_metrics and self._should_collect_customer_sdkstats():
-                                    _track_dropped_items(self._customer_sdkstats_metrics, envelopes, DropCode.CLIENT_EXCEPTION, "Error parsing redirect information.")
+                                    _track_dropped_items(self._customer_sdkstats_metrics, envelopes, DropCode.CLIENT_EXCEPTION, _CLIENT_EXCEPTION)
                                 logger.error(
                                     "Error parsing redirect information.",
                                 )
@@ -354,7 +355,7 @@ class BaseExporter:
                                     self._customer_sdkstats_metrics,
                                     envelopes,
                                     DropCode.CLIENT_EXCEPTION,
-                                    "Error sending telemetry because of circular redirects. Please check the integrity of your connection string."
+                                    _CLIENT_EXCEPTION
                                 )
                             logger.error(
                                 "Error sending telemetry because of circular redirects. "
@@ -413,7 +414,7 @@ class BaseExporter:
 
                 # Track dropped items in customer sdkstats for general exceptions
                 if self._customer_sdkstats_metrics and self._should_collect_customer_sdkstats():
-                    _track_dropped_items(self._customer_sdkstats_metrics, envelopes, DropCode.CLIENT_EXCEPTION, str(ex))
+                    _track_dropped_items(self._customer_sdkstats_metrics, envelopes, DropCode.CLIENT_EXCEPTION, _CLIENT_EXCEPTION)
 
                 if self._should_collect_stats():
                     _update_requests_map(_REQ_EXCEPTION_NAME[1], value=ex.__class__.__name__)