+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/statsbeat/_customer_sdkstats.py
@@ -211,28 +211,28 @@ class CustomerSdkStatsMetrics(metaclass=Singleton): # pylint: disable=too-many-i
             return categorize_status_code(drop_code)
 
         if drop_code == DropCode.CLIENT_EXCEPTION:
-            return exception_message if exception_message else "unknown_exception"
+            return exception_message if exception_message else "Client exception"
 
         drop_code_reasons = {
-            DropCode.CLIENT_READONLY: "readonly_mode",
-            DropCode.CLIENT_STORAGE_DISABLED: "local storage is disabled",
-            DropCode.CLIENT_PERSISTENCE_CAPACITY: "persistence_full",
+            DropCode.CLIENT_READONLY: "Readonly mode",
+            DropCode.CLIENT_STORAGE_DISABLED: "Local storage is disabled",
+            DropCode.CLIENT_PERSISTENCE_CAPACITY: "Persistence full",
         }
 
-        return drop_code_reasons.get(drop_code, "unknown_reason")
+        return drop_code_reasons.get(drop_code, "Unknown reason")
 
     def _get_retry_reason(self, retry_code: RetryCodeType, exception_message: Optional[str] = None) -> str:
         if isinstance(retry_code, int):
             return categorize_status_code(retry_code)
 
         if retry_code == RetryCode.CLIENT_EXCEPTION:
-            return exception_message if exception_message else "unknown_exception"
+            return exception_message if exception_message else "Client exception"
 
         retry_code_reasons = {
-            RetryCode.CLIENT_TIMEOUT: "client_timeout",
-            RetryCode.UNKNOWN: "unknown_reason",
+            RetryCode.CLIENT_TIMEOUT: "Client timeout",
+            RetryCode.UNKNOWN: "Unknown reason",
         }
-        return retry_code_reasons.get(retry_code, "unknown_reason")
+        return retry_code_reasons.get(retry_code, "Unknown reason")
 
 # Global customer sdkstats singleton
 _CUSTOMER_SDKSTATS_METRICS = None